+++ cypress/integration/executions_spec.js
@@ -311,9 +311,14 @@ describe('Dashboard Executions Page', () => {
           cy.contains('State Machine Arn').next().should('have.text', stateMachine);
           cy.contains('Started').next().should('have.text', startMatch);
           cy.contains('Ended').next().should('have.text', endMatch);
-          cy.contains('Granule ID').next().should('have.text', granuleId);
-          cy.contains('Associated Executions List').next().should('have.text', 'Link');
+          cy.contains('Parent Workflow Execution').next().should('have.text', 'N/A');
         });
+
+      cy.get('.table .thead .th').eq(0).contains('Granule ID');
+      cy.get('.table .thead .th').eq(1).contains('Associated Executions List');
+
+      cy.get('.table .tbody .td').eq(0).contains(granuleId);
+      cy.get('.table .tbody .td').eq(1).contains('Link');
 
       cy.getFakeApiFixture('executions').as('executionsFixture');
       cy.get('@executionsFixture').its('results')
@@ -351,15 +356,11 @@ describe('Dashboard Executions Page', () => {
       cy.get('@response').then((resp) => {
         cy.wrap(resp.response.body.execution.granules[0]).as('granule');
         cy.get('@granule').then((granule) => {
-          cy.get('.status--process').within(() => {
-            cy.contains('Associated Executions List')
-              .next()
-              .should('have.text', 'Link')
-              .children('a')
-              .should('have.attr', 'href')
-              .should('include', `${granule.collectionId}/${granule.granuleId}`);
-            cy.contains('Associated Executions List').next().children('a').click();
-          });
+          cy.get('.table .tbody .td').eq(0).contains(granule.granuleId);
+          cy.get('.table .tbody .td').eq(1).should('have.text', 'Link').children('a')
+            .should('have.attr', 'href')
+            .should('include', `${granule.collectionId}/${granule.granuleId}`);
+          cy.get('.table .tbody .td').eq(1).children('a').click();
         });
       });